requirements.txt
-backend

asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv

-frontend
axios
react-router-dom 
jwt-decode
shadcn